% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothn.R
\name{dctn}
\alias{dctn}
\alias{idctn}
\title{Multidimensional Discrete Cosine Transform along one or more axes}
\usage{
dctn(x, d = NULL)

idctn(x, d = NULL)
}
\arguments{
\item{x}{Array. A n-dimensional array.}

\item{d}{Integer. A vector of dimensions for applying the 1-D DCT/IDCT. If
\code{d = NULL}, all dimensions will be used.}
}
\value{
An array of the same shape as \code{x}. If \code{x} is a vector, the
output will also be a vector.
}
\description{
This function uses the seperability property of n-D Discrete Consine
Transform (DCT) and apply the one dimensional DCT/IDCT along each specified
axis. See \code{\link[gsignal:dct]{gsignal::dct()}} and \code{\link[gsignal:dct2]{gsignal::dct2()}} for 1-D DCT and 2-D DCT
respectively. See \code{\link[gsignal:idct]{gsignal::idct()}} and \code{\link[gsignal:idct2]{gsignal::idct2()}} for
1-D IDCT and 2-D IDCT respectively.
}
\section{Functions}{
\itemize{
\item \code{dctn()}: Discrete Cosine Transform

\item \code{idctn()}: Inverse Discrete Cosine Transform

}}
\examples{

# Apply DCT for rows and columns
dctn(cars)

# Apply DCT for rows
dctn(cars, 2)

# Reconstruct `cars`
idctn(dctn(cars))

}
