% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothn.R
\name{replace_d}
\alias{replace_d}
\title{Replace elements in vectors of array along an axis}
\usage{
replace_d(x, d, fn, ...)
}
\arguments{
\item{x}{Array. A n-dimensional array.}

\item{d}{Integer. The target dimension. e.g. 1 for slicing and replacing
along row, 2 for slicing and replacing along column.}

\item{fn}{Function. A function that takes an original vector as input and
outputs a new vector of the same length.}

\item{...}{Additional arguments passed to \code{fn}.}
}
\value{
An array of the same shape as \code{x}.
}
\description{
This function replaces elements in vectors of an array along an axis.
See \code{\link[=slice_d]{slice_d()}} for more details about slicing vectors along an axis.
}
\examples{

# Replace columns with column sums
replace_d(as.matrix(cars), d = 1, sum)

# Replace rows with row sums
replace_d(as.matrix(cars), d = 2, sum)
}
