% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothn.R
\name{slice_d}
\alias{slice_d}
\title{Get indices of vectors of array along an axis}
\usage{
slice_d(x, d, get_data = FALSE)
}
\arguments{
\item{x}{Array. A n-dimensional array.}

\item{d}{Integer. The target dimension. e.g. 1 for slicing along row, 2 for
slicing along column.}

\item{get_data}{Boolean. Whether to retrieve data using the indices.}
}
\value{
A tibble with two or three columns. The first column is \code{vector_id}
providing a vector of unique IDs for each vector. The second column is
\code{index} providing a list of matrices for accessing each vector. The third
column is \code{data} if \code{get_data = TRUE} providing a list of vectors.
}
\description{
This function slices all vectors of an array along an axis.
}
\examples{

# Slice along row. For 2D array, returns all columns.
slice_d(as.matrix(cars), d = 1)

# Slice along column. For 2D array, returns all rows.
slice_d(as.matrix(cars), d = 2)

# Slice and access.
slice_d(as.matrix(cars), d = 1, get_data = TRUE)

}
